+++ CHANGELOG.md
@@ -170,6 +170,9 @@ bearing with us as we move towards our first stable Lightning release.)
   [#1348](https://github.com/OpenFn/Lightning/issues/1348)
 - Modify CompleteRun to return error changeset when run not found
   [#1393](https://github.com/OpenFn/Lightning/issues/1393)
+- Drop invocation reasons from DB
+  [#1412](https://github.com/OpenFn/Lightning/issues/1412)
+
 
 ## [v0.9.3] - 2023-09-27